Moving around and delete some stray code
